Change to scoped and only get env once
