+++ sdk/tests/CMakeLists.txt
@@ -33,7 +33,7 @@ add_executable(hipdnn_sdk_tests
     test_utilities/TestCpuFpReferenceValidation.cpp
     test_utilities/TestCpuFpReferenceMiopenRmsValidation.cpp
     test_utilities/TestTempDirectory.cpp
-    test_utilities/TestEnvironmentVariableGuard.cpp
+    test_utilities/TestScopedEnvironmentVariableSetter.cpp
     test_utilities/TestLoggingUtils.cpp
 )
